+++ SquidSquAds/src/main/java/com/squidsquads/repository/stats/BrowserTypesItemRepository.java
@@ -6,5 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface BrowserTypesItemRepository extends JpaRepository<BrowserTypesItem, Long> {
-    List<BrowserTypesItem> findAllByWebsiteIDOrderByRatioDesc(Long websiteID);
+
+    List<BrowserTypesItem> findAllByWebsiteIDOrderByRatioDesc(Integer websiteID);
+
 }
